itest: move `testDataLossProtection` to backup test file
